refactor: cambiar definición de modelo
